Refactor that it is possible to declare paths inside the 'gitlab_db_sensor.py' and 'gitlab_db_mdgen.py' scripts for the generated files
Previously the '_generated' directories where created relative to the corresponding scripts.
Now it is possible to declare them as function argument and therefore move all generated files inside one
"_generated" directory with sub directories declared in the main script. This "_generated" directory is located at the root directory
of the repository.