Add workaround for a rdflib bug where a "BASE" directive isn't used by the serializer if explicitly given in the function call
The serializer uses the optional "base" property of the graph object (points to None) that doesn't
get overwritten in this occasion by specifying it in the serialize(...) function
call as it should be.
This is a bug in the rdflib python package.

Therefore a optional initialization argument 'base' is added to the "Kraken"
class to be able to explicitly specify the "BASE" directive while initializing
the graph object.